feat(website): offline banner support
